More v6 work
