+++ lib/index.js
@@ -1385,19 +1385,6 @@ function validateShard(shard) {
 		shard[1] > 1
 	) ? shard : null;
 }
-function identifyOrResume(client) {
-	var payload, internals = client.internals;
-
-	if (internals.sequence && internals.token && internals.sessionID) {
-		payload = Payloads.RESUME(client);
-	} else {
-		payload = Payloads.IDENTIFY(client);
-		if (client._shard) payload.d.shard = client._shard;
-	}
-	client._connecting = false;
-
-	return ws_send(client._ws, payload);
-}
 
 /* - Functions - Websocket Handling - */
 function handleWSMessage(data, flags) {
@@ -2295,121 +2282,5 @@ Object.keys(Discord.Permissions).forEach(function(pn) {
 		enumerable: true
 	});
 });
-
-/* Payloads */
-(function() {
-	Payloads = {
-		IDENTIFY: function(client) {
-			return {
-				op: 2,
-				d: {
-					token: client.internals.token,
-					v: GATEWAY_VERSION,
-					compress: isNode && !!Zlib.inflateSync,
-					large_threshold: LARGE_THRESHOLD,
-					properties: {
-						$os: isNode ? require('os').platform() : navigator.platform,
-						$browser:"discord.io",
-						$device:"discord.io",
-						$referrer:"",
-						$referring_domain:""
-					}
-				}
-			};
-		},
-		RESUME: function(client) {
-			return {
-				op: 6,
-				d: {
-					seq: client.internals.s,
-					token: client.internals.token,
-					session_id: client.internals.sessionID
-				}
-			};
-		},
-		HEARTBEAT: function(client) {
-			return {op: 1, d: client.internals.sequence};
-		},
-		ALL_USERS: function(client) {
-			return {op: 12, d: Object.keys(client.servers)};
-		},
-		STATUS: function(status_string, game_object) {
-			return {
-				op: 3,
-				d: {
-					status: status_string,
-					game: game_object
-				},
-				afk: false,
-				since: 0
-			}
-		},
-		STATUS: function(client, input) {
-			return {
-				op: 3,
-				d: {
-					status: input.status || client.status,
-					game: type(input.game) === 'object' ?
-						{
-							name: input.game.name ? String(input.game.name) : null,
-							type: input.game.type ? Number(input.game.type) : null,
-							url: input.game.url ? String(input.game.url) : null
-						} :
-						null,
-					afk: false,
-					since: 0
-				}
-			}
-		},
-		UPDATE_VOICE: function(serverID, channelID) {
-			return {
-				op: 4,
-				d: {
-					guild_id: serverID,
-					channel_id: channelID,
-					self_mute: false,
-					self_deaf: false
-				}
-			};
-		},
-		OFFLINE_USERS: function(array) {
-			return {
-				op: 8,
-				d: {
-					guild_id: array.splice(0, 50),
-					query: "",
-					limit: 0
-				}
-			};
-		},
-		VOICE_SPEAK: function(v) {
-			return {op:5, d:{ speaking: !!v, delay: 0 }};
-		},
-		VOICE_IDENTIFY: function(clientID, voiceSession) {
-			return {
-				op: 0,
-				d: {
-					server_id: voiceSession.serverID,
-					user_id: clientID,
-					session_id: voiceSession.session,
-					token: voiceSession.token
-				}
-			};
-		},
-		VOICE_DISCOVERY: function(ip, port, mode) {
-			return {
-				op:1,
-				d:{
-					protocol:"udp",
-					data:{
-						address: ip,
-						port: Number(port),
-						mode: mode
-					}
-				}
-			};
-		}
-	}
-})();
 
 })(typeof exports === 'undefined'? this.Discord = {} : exports);
