+++ lib/new_index.js
@@ -10,7 +10,7 @@ var CURRENT_VERSION      = "3.0.0",
 	LARGE_THRESHOLD      = 250,
 	MAX_REQUESTS         = 5,
 	CONNECT_WHEN         = null,
-	Endpoints, Payloads,
+	Endpoints,
 	Opcodes = {
 		Gateway: {
 			Dispatch: 0,
@@ -27,7 +27,119 @@ var CURRENT_VERSION      = "3.0.0",
 			Heartbeat_ACK: 11
 		},
 		Voice: {}
-	};
+	},
+	Payloads = {
+		IDENTIFY: function(client) {
+			return {
+				op: 2,
+				d: {
+					token: client.internals.token,
+					v: WS_GATEWAY_VERSION,
+					compress: is_node && !!zlib.inflateSync,
+					large_threshold: LARGE_THRESHOLD,
+					properties: {
+						$os: is_node ? require('os').platform() : navigator.platform,
+						$browser:"discord.io",
+						$device:"discord.io",
+						$referrer:"",
+						$referring_domain:""
+					}
+				}
+			};
+		},
+		RESUME: function(client) {
+			return {
+				op: 6,
+				d: {
+					seq: client.internals.s,
+					token: client.internals.token,
+					session_id: client.internals.sessionID
+				}
+			};
+		},
+		HEARTBEAT: function(client) {
+			return {op: 1, d: client.internals.sequence};
+		},
+		ALL_USERS: function(client) {
+			return {op: 12, d: Object.keys(client.servers)};
+		},
+		STATUS: function(status_string, game_object) {
+			return {
+				op: 3,
+				d: {
+					status: status_string,
+					game: game_object
+				},
+				afk: false,
+				since: 0
+			}
+		},
+		STATUS: function(client, input) {
+			return {
+				op: 3,
+				d: {
+					status: input.status || client.status,
+					game: type(input.game) === 'object' ?
+						{
+							name: input.game.name ? String(input.game.name) : null,
+							type: input.game.type ? Number(input.game.type) : null,
+							url: input.game.url ? String(input.game.url) : null
+						} :
+						null,
+					afk: false,
+					since: 0
+				}
+			}
+		},
+		UPDATE_VOICE: function(serverID, channelID) {
+			return {
+				op: 4,
+				d: {
+					guild_id: serverID,
+					channel_id: channelID,
+					self_mute: false,
+					self_deaf: false
+				}
+			};
+		},
+		OFFLINE_USERS: function(array) {
+			return {
+				op: 8,
+				d: {
+					guild_id: array.splice(0, 50),
+					query: "",
+					limit: 0
+				}
+			};
+		},
+		VOICE_SPEAK: function(v) {
+			return {op:5, d:{ speaking: !!v, delay: 0 }};
+		},
+		VOICE_IDENTIFY: function(clientID, voiceSession) {
+			return {
+				op: 0,
+				d: {
+					server_id: voiceSession.serverID,
+					user_id: clientID,
+					session_id: voiceSession.session,
+					token: voiceSession.token
+				}
+			};
+		},
+		VOICE_DISCOVERY: function(ip, port, mode) {
+			return {
+				op:1,
+				d:{
+					protocol:"udp",
+					data:{
+						address: ip,
+						port: Number(port),
+						mode: mode
+					}
+				}
+			};
+		}
+	}
 
 if (is_node) {
 	udp         = require('dgram'),
@@ -360,7 +472,7 @@ class Requester {
 						.once('end', function() {
 							chunks = Buffer.concat(chunks);
 							//[Object Headers, Object Body]
-							try { chunks = Zlib.gunzipSync(chunks); } catch(e) {}
+							try { chunks = zlib.gunzipSync(chunks); } catch(e) {}
 							return resolve([res.headers, JSON.parse(chunks)]);
 						});
 				});
@@ -504,33 +616,65 @@ async function start_connection(client, gateway = Endpoints.WS_GATEWAY) {
 }
 
 async function handle_ws_message(client, event) {
-	var message = decompress_ws_message(event.data, event.binary);
+	var message, payload;
+
+	message = decompress_ws_message(event.data, event.binary);
+console.log(message);
 	switch(message.op) {
 		case Opcodes.Gateway.Dispatch:
 			client.internals.sequence = message.s;
-			await handle_dispatch(client, message);
+			handle_dispatch(client, message);
 			break;
 		case Opcodes.Gateway.Reconnect:
 			break;
 		case Opcodes.Gateway.Invalid_Session:
 			break;
 		case Opcodes.Gateway.Hello:
+			payload = identify_or_resume(client);
+			ws_send(client._ws, payload);
+
+			client._main_keep_alive = setInterval(
+				keep_alive.bind(null, client),
+				message.d.heartbeat_interval
+			);
 			break;
 		case Opcodes.Gateway.Heartbeat_ACK: 
 			break;
 	}
 }
 
-async function handle_dispatch(client, message) {
+function handle_dispatch(client, message) {
+	console.log(message);
 	switch (message.t) {
 		case "READY":
-			console.log(message);
 			break;
 	}
 }
 
 function decompress_ws_message(message, binary) {
 	return binary ? JSON.parse(zlib.inflateSync(message).toString()) : JSON.parse(message);
+}
+
+function identify_or_resume(client) {
+	var payload, internals = client.internals;
+
+	if (internals.sequence && internals.token && internals.sesionID) {
+		return Payloads.RESUME(client);
+	}
+
+	payload = Payloads.IDENTIFY(client);
+	if (client._shard) payload.d.shard = client._shard;
+
+	return payload;
+}
+
+function keep_alive(client) {
+	client.internals.heartbeat = setTimeout(
+		client._ws.close.bind(client._ws, 1e3, 'No heartbeat received'),
+		15e3
+	);
+	client.internals._lastHB = Date.now();
+	ws_send(client._ws, Payloads.HEARTBEAT(client));
 }
 
 /**
